move config package (coreth c167754ba)
