+++ plugin/evm/config/config_test.go
@@ -1,7 +1,7 @@
 // (c) 2019-2020, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package evm
+package config
 
 import (
 	"encoding/json"
